split CometShuffleExternalSorter into sync/async implementations
